Move tests related to creation into its respective file
